+++ test/unit/mongo_db_restore/main.py
@@ -42,7 +42,7 @@ class ProgramLock(object):
     Description:  Class stub holder for gen_class.ProgramLock class.
 
     Methods:
-        __init__ -> Class initialization.
+        __init__
 
     """
 
@@ -53,8 +53,6 @@ class ProgramLock(object):
         Description:  Class initialization.
 
         Arguments:
-            (input) cmdline -> Argv command line.
-            (input) flavor -> Lock flavor ID.
 
         """
 
@@ -68,22 +66,18 @@ class UnitTest(unittest.TestCase):
 
     Description:  Class which is a representation of a unit testing.
 
-    Super-Class:  unittest.TestCase
-
-    Sub-Classes:
-
     Methods:
-        setUp -> Initialize testing environment.
-        test_help_true -> Test help if returns true.
-        test_help_false -> Test help if returns false.
-        test_arg_req_true -> Test arg_require if returns true.
-        test_arg_req_false -> Test arg_require if returns false.
-        test_arg_dir_chk_crt_true -> Test arg_dir_chk_crt if returns true.
-        test_arg_dir_chk_crt_false -> Test arg_dir_chk_crt if returns false.
-        test_run_program -> Test run_program function.
-        test_programlock_id -> Test with ProgramLock with flavor id.
-        test_programlock_false -> Test with ProgramLock returns False.
-        test_programlock_true -> Test with ProgramLock returns True.
+        setUp
+        test_help_true
+        test_help_false
+        test_arg_req_true
+        test_arg_req_false
+        test_arg_dir_chk_crt_true
+        test_arg_dir_chk_crt_false
+        test_run_program
+        test_programlock_id
+        test_programlock_false
+        test_programlock_true
 
     """