+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 
 ## [0.0.2] - 2021-12-22
+### Changed
+- Documentation updates.
 
 
 ## [0.0.1] - 2020-10-29